File reorganisation
Moved the ducky generation into their own package
